arregla estilo de espaciado

+++ 4.ps1
@@ -1,8 +1,6 @@
 ####################################################
 #######        CERRAR PROCESOS    ##############
 ####################################################
-$a=read-host -prompt "Introduce ID del proceso"
-$b=read-host -prompt "Seguro que quieres cerrar el proceso(s/n)?"
-if($b -eq "s"){
-Stop-Process $a
-}
+$a = read-host -prompt "Introduce ID del proceso"
+$b = read-host -prompt "Seguro que quieres cerrar el proceso(s/n)?"
+if( $b -eq "s" ){ Stop-Process $a }
